Clean up the main App.tsx file a bit.
